Fixed show/hide button UI
