+++ common/src/main/java/io/novafoundation/nova/common/view/ExpandableView.kt
@@ -6,7 +6,6 @@ import android.graphics.Rect
 import android.util.AttributeSet
 import android.view.View
 import android.view.animation.AccelerateDecelerateInterpolator
-import android.widget.TextView
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.animation.doOnEnd
 import androidx.core.animation.doOnStart
@@ -15,7 +14,6 @@ import io.novafoundation.nova.common.R
 import io.novafoundation.nova.common.utils.getResourceIdOrNull
 import io.novafoundation.nova.common.utils.makeGone
 import io.novafoundation.nova.common.utils.makeVisible
-import io.novafoundation.nova.common.utils.setDrawableEnd
 
 enum class ExpandableViewState {
     COLLAPSED,
@@ -163,23 +161,4 @@ class ExpandableView @JvmOverloads constructor(
     }
 
     private fun findViewByIdOrNull(id: Int?): View? = id?.let { findViewById(it) }
-}
-
-fun ExpandableView.bindWithHideShowButton(hideShowButton: TextView) {
-    hideShowButton.setHideShowButtonState(currentState())
-    setCallback { hideShowButton.setHideShowButtonState(it) }
-}
-
-private fun TextView.setHideShowButtonState(state: ExpandableViewState) {
-    when (state) {
-        ExpandableViewState.COLLAPSED -> {
-            text = context.getString(R.string.common_show)
-            setDrawableEnd(R.drawable.ic_eye_show, widthInDp = 20, paddingInDp = 8)
-        }
-
-        ExpandableViewState.EXPANDED -> {
-            text = context.getString(R.string.common_hide)
-            setDrawableEnd(R.drawable.ic_eye_hide, widthInDp = 20, paddingInDp = 8)
-        }
-    }
 }
