+++ feature-multisig/operations/src/main/res/layout/fragment_multisig_operation_details.xml
@@ -65,8 +65,8 @@
             </io.novafoundation.nova.common.view.TableView>
 
             <io.novafoundation.nova.common.view.TableView
-                android:layout_width="match_parent"
                 android:id="@+id/multisigPendingOperationDetailsCallTable"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="8dp" />
 
@@ -95,6 +95,7 @@
                 android:layout_marginTop="8dp"
                 android:background="@drawable/bg_block_12"
                 android:paddingTop="12dp"
+                app:chevronId="@+id/multisigOperationShowHideButtonIcon"
                 app:expandableId="@+id/multisigOperationSignatories"
                 app:supportAnimation="false">
 
@@ -115,21 +116,29 @@
                     android:id="@+id/multisigOperationShowHideButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginEnd="16dp"
                     android:gravity="center"
                     android:includeFontPadding="false"
+                    android:paddingEnd="4dp"
                     android:textColor="@color/text_secondary"
-                    app:drawableTint="@color/icon_secondary"
+                    app:layout_constraintBottom_toBottomOf="@+id/multisigOperationShowHideButtonIcon"
+                    app:layout_constraintEnd_toStartOf="@+id/multisigOperationShowHideButtonIcon"
+                    app:layout_constraintTop_toTopOf="parent"
+                    tools:text="Show" />
+
+                <ImageView
+                    android:id="@+id/multisigOperationShowHideButtonIcon"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="16dp"
+                    android:src="@drawable/ic_chevron_up"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintTop_toTopOf="parent"
-                    tools:drawableEnd="@drawable/ic_eye_show"
-                    tools:drawablePadding="8dp"
-                    tools:text="Show" />
+                    app:tint="@color/icon_secondary" />
 
                 <Space
                     android:layout_width="match_parent"
                     android:layout_height="12dp"
-                    app:layout_constraintTop_toBottomOf="@+id/multisigOperationShowHideButton" />
+                    app:layout_constraintTop_toBottomOf="@+id/multisigOperationShowHideButtonIcon" />
 
                 <androidx.recyclerview.widget.RecyclerView
                     android:id="@+id/multisigOperationSignatories"
@@ -139,7 +148,7 @@
                     android:overScrollMode="never"
                     android:paddingBottom="8dp"
                     app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                    app:layout_constraintTop_toBottomOf="@+id/multisigOperationShowHideButton"
+                    app:layout_constraintTop_toBottomOf="@+id/multisigOperationShowHideButtonIcon"
                     tools:itemCount="4"
                     tools:listitem="@layout/item_multisig_signatory_account" />
